test: extract RefDataset splice settings tests to unit/splice/
